Cleanup in number theory
Remove or hide deprecated functions

Consolidate some source files
