+++ src/lib/pubkey/ec_group/curve_gfp.cpp
@@ -12,6 +12,7 @@
 #include <botan/reducer.h>
 #include <botan/internal/mp_core.h>
 #include <botan/internal/mp_asmi.h>
+#include <botan/internal/monty.h>
 
 namespace Botan {
